requirements.txt
bonderizer @ git+https://github.com/twonds/bonderizer-py@main
zope.interface==5.5.2
